+++ TransitCity/Utility/Units/Acceleration.cs
@@ -2,14 +2,14 @@
 {
     public class Acceleration
     {
-        private readonly float _metersPerSecondsSquared;
+        private readonly double _metersPerSecondsSquared;
 
-        internal Acceleration(float metersPerSecondSquared)
+        internal Acceleration(double metersPerSecondSquared)
         {
             _metersPerSecondsSquared = metersPerSecondSquared;
         }
 
-        public float MetersPerSecondSquared => _metersPerSecondsSquared;
+        public double MetersPerSecondSquared => _metersPerSecondsSquared;
 
         public override string ToString()
         {
@@ -20,11 +20,11 @@
 
         public static Acceleration operator -(Acceleration a1, Acceleration a2) => new Acceleration(a1.MetersPerSecondSquared - a2.MetersPerSecondSquared);
 
-        public static Acceleration operator *(Acceleration a, float f) => new Acceleration(a.MetersPerSecondSquared * f);
+        public static Acceleration operator *(Acceleration a, double f) => new Acceleration(a.MetersPerSecondSquared * f);
 
-        public static Acceleration operator *(float f, Acceleration a) => a * f;
+        public static Acceleration operator *(double f, Acceleration a) => a * f;
 
-        public static Acceleration operator /(Acceleration a, float f) => new Acceleration(a.MetersPerSecondSquared / f);
+        public static Acceleration operator /(Acceleration a, double f) => new Acceleration(a.MetersPerSecondSquared / f);
 
         public static Speed operator *(Acceleration a, Duration t) => new Speed(a.MetersPerSecondSquared * t.Seconds);
 
@@ -38,6 +38,6 @@
 
         public static bool operator >=(Acceleration a1, Acceleration a2) => a1.MetersPerSecondSquared >= a2.MetersPerSecondSquared;
 
-        public static Acceleration FromMetersPerSecondSquared(float metersPerSecondSquared) => new Acceleration(metersPerSecondSquared);
+        public static Acceleration FromMetersPerSecondSquared(double metersPerSecondSquared) => new Acceleration(metersPerSecondSquared);
     }
 }
