+++ TransitCity/Utility/Units/Speed.cs
@@ -2,16 +2,16 @@
 {
     public class Speed
     {
-        private readonly float _metersPerSecond;
+        private readonly double _metersPerSecond;
 
-        internal Speed(float metersPerSecond)
+        internal Speed(double metersPerSecond)
         {
             _metersPerSecond = metersPerSecond;
         }
 
-        public float MetersPerSecond => _metersPerSecond;
+        public double MetersPerSecond => _metersPerSecond;
 
-        public float KilometersPerHour => _metersPerSecond * 3.6f;
+        public double KilometersPerHour => _metersPerSecond * 3.6;
 
         public override string ToString()
         {
@@ -22,11 +22,11 @@
 
         public static Speed operator -(Speed s1, Speed s2) => new Speed(s1.MetersPerSecond - s2.MetersPerSecond);
 
-        public static Speed operator *(Speed s, float f) => new Speed(s.MetersPerSecond * f);
+        public static Speed operator *(Speed s, double f) => new Speed(s.MetersPerSecond * f);
 
-        public static Speed operator *(float f, Speed s) => s * f;
+        public static Speed operator *(double f, Speed s) => s * f;
 
-        public static Speed operator /(Speed s, float f) => new Speed(s.MetersPerSecond / f);
+        public static Speed operator /(Speed s, double f) => new Speed(s.MetersPerSecond / f);
 
         public static Distance operator *(Speed s, Duration t) => new Distance(s.MetersPerSecond * t.Seconds);
 
@@ -40,8 +40,8 @@
 
         public static bool operator >=(Speed s1, Speed s2) => s1.MetersPerSecond >= s2.MetersPerSecond;
 
-        public static Speed FromMetersPerSecond(float metersPerSecond) => new Speed(metersPerSecond);
+        public static Speed FromMetersPerSecond(double metersPerSecond) => new Speed(metersPerSecond);
 
-        public static Speed FromKilometersPerHour(float kilometersPerHour) => new Speed(kilometersPerHour / 3.6f);
+        public static Speed FromKilometersPerHour(double kilometersPerHour) => new Speed(kilometersPerHour / 3.6);
     }
 }
